:recycle: Refactoring all endpoints to api version 1
Moving files, create v1 folder, rename routes and update swagger documentation.
