+++ src/api/v1/docs/swagger.yaml
@@ -12,8 +12,8 @@ externalDocs:
   description: Find out more about Proyect
   url: https://www.notion.so/luissolano/Proyecto-base-MERN-Stack-0db74dd9fbe549239d6ba0f61e877a49
 servers:
-  - url: http://localhost:3000/api
-  - url: https://base-node-ts.up.railway.app/api
+  - url: http://localhost:3000/api/v1
+  - url: https://base-node-ts.up.railway.app/api/v1
 
 tags:
   - name: Auth
